% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.plot.R
\name{surv.plot}
\alias{surv.plot}
\title{Publication Ready Kaplan-Meier Plot}
\usage{
surv.plot(
  fit,
  mark.censoring = TRUE,
  conf.int = fit$conf.int,
  conf.band = TRUE,
  conf.line = FALSE,
  conf.band.col = col,
  conf.band.transparent = 0.25,
  conf.type = "log-log",
  grid = FALSE,
  col = NULL,
  main = NULL,
  sub = NULL,
  xlab = "Time",
  ylab = "Estimated survival probability",
  cex = NULL,
  cex.lab = 1,
  cex.axis = 1,
  bty = "l",
  lty = c("solid", "dotted", "dotted"),
  lwd = 1,
  xlim = seq(from = 0, to = ceiling(max(fit$time)) + ceiling(min(fit$time))),
  ylim = seq(from = 0, to = 1, by = 0.25),
  show.legend = TRUE,
  legend.position = "topright",
  legend.name = NULL,
  legend.text.font = 1,
  legend.cex = 1,
  legend.title = NULL,
  legend.title.cex = 1,
  segment.type = 3,
  segment.timepoint = NULL,
  segment.quantile = NULL,
  segment.main = NULL,
  segment.annotation = "right",
  segment.col = "#666666",
  segment.annotation.col = col,
  segment.lty = "dashed",
  segment.lwd = 1,
  segment.cex = 1,
  segment.annotation.space = 0.06,
  segment.font = 1,
  segment.main.font = 1,
  stat = "none",
  stat.position = "bottomleft",
  stat.col = "black",
  stat.cex = 1,
  stat.font = 1,
  risktable = TRUE,
  risktable.axislab.pos = 2.5,
  risktable.margin.bottom = 5,
  risktable.margin.left = 6.5,
  risktable.title = "Number at risk",
  risktable.title.font = 2,
  risktable.title.col = "black",
  risktable.title.position = par("usr")[1] - (par("usr")[2] - par("usr")[1]) * 0.15,
  risktable.cex = 1,
  risktable.title.cex = 1,
  risktable.name.cex = 1,
  risktable.col = "black",
  risktable.name.font = 1,
  risktable.name.col = "#666666",
  risktable.name.position = par("usr")[1] - (par("usr")[2] - par("usr")[1]) * 0.15
)
}
\arguments{
\item{fit}{An object of class \code{survfit} containing survival data.}

\item{mark.censoring}{Mark censoring events on the curves. Logical (default: TRUE)}

\item{conf.int}{Display confidence intervals. (FALSE, TRUE for 95\% confidence intervals, or specify a numeric value for desired coverage.}

\item{conf.band}{Mapping the specified coverage probability}

\item{conf.line}{A logical value for drawing confidence line (Default: FALSE).}

\item{conf.band.col}{Colour(s) for confidence band. Can accept a single value for colour, or a vector of colour values.}

\item{conf.band.transparent}{Transparency for the confidence band.}

\item{conf.type}{Transformation type of the confidence interval.Options: "log", "log-log", "plain", "logit", "arcsin";(default: "log-log").}

\item{grid}{Draw a grid on the plot Logical (default: FALSE)}

\item{col}{Colour(s) for the survival curves. Can accept a single value for colour, or a vector of colour values to set colour(s).}

\item{main}{Title of the plot.}

\item{sub}{Subtitle of the plot.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{cex}{A numeric value specifying the global size of the text.}

\item{cex.lab}{A numeric value specifying the size of the xlab and ylab text.}

\item{cex.axis}{A numeric value specifying the size of the axis size.}

\item{bty}{The type of box to be drawn around the plot ("n","o","7","L","C","U")}

\item{lty}{A vector of string specifying line types for each curve (“blank”, “solid”, “dashed”, “dotted”, “dotdash”, “longdash”, “twodash”).}

\item{lwd}{A vector of numeric values for line widths.}

\item{xlim}{X-axis limits specified as a sequence; seq(starting value,  end value, number of increment of the sequence).}

\item{ylim}{Y-axis limits specified as a sequence; seq(starting value,  end value, number of increment of the sequence).}

\item{show.legend}{Display legend.}

\item{legend.position}{Position of the legend (c(x,y), "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center").}

\item{legend.name}{A vector of string specifying the name(s) of stratum.}

\item{legend.text.font}{An integer specifying the font style of the legend text; (1: normal, 2: bold, 3: italic, 4: bold and italic).}

\item{legend.cex}{Expansion factor for legend text.}

\item{legend.title}{The title of the legend.}

\item{legend.title.cex}{Expansion factor for legend title.}

\item{segment.type}{A numeric value specifying the layout of the segment (1: Draws specified segment (full bandwidth), 2: Draws specified segment, 3: Drawing vertical and horizontal segment).}

\item{segment.timepoint}{A single value or a vector of fixed time points of segment(s).}

\item{segment.quantile}{A single value or a vector of fixed quantile of segment(s) at a fixed quantile (e.g. 0.5 corresponds to median).}

\item{segment.main}{Title of segment text.}

\item{segment.annotation}{Position of the segment annotation: (c(x,y), "bottomleft", "left", "right", "none").}

\item{segment.col}{Can accept a single value for colour, or a vector of colour values to set colour(s).}

\item{segment.annotation.col}{Can accept a single value for colour, or a vector of colour values to set colour(s).}

\item{segment.lty}{A vector of string specifying line types for each curve (“blank”, “solid”, “dashed”, “dotted”, “dotdash”, “longdash”, “twodash”).}

\item{segment.lwd}{A vector of numeric values for line widths.}

\item{segment.cex}{A numeric values specifying the size of the segment annotation size.}

\item{segment.annotation.space}{Spacing between the text in unit of x-coordinates.}

\item{segment.font}{A numeric value specifying the font face (1 = plain, 2 = bold, 3 = italic, 4 = bold-italic, ...).}

\item{segment.main.font}{A numeric value specifying the fon face (1 = plain, 2 = bold, 3 = italic, 4 = bold-italic, ...).}

\item{stat}{Statistics which is displayed in the plot ("logrank", "coxph", "coxmodel", "none").}

\item{stat.position}{Position where the stat should be displayed: (c(x,y), "bottomleft", "left", "right", "none").}

\item{stat.col}{Can accept a single value for colour.}

\item{stat.cex}{A numeric value specifying the size of the stat size.}

\item{stat.font}{The font face (1 = plain, 2 = bold, 3 = italic, 4 = bold-italic).}

\item{risktable}{A logical value for drawing risk table (Default: TRUE).}

\item{risktable.axislab.pos}{Position of the X and Y label, specified on which line.}

\item{risktable.margin.bottom}{Modify bottom margin of the plot region in line unit (Default: 5).}

\item{risktable.margin.left}{Modify left margin of the plot region in line unit (Default: 7).}

\item{risktable.title}{Title of risk table.}

\item{risktable.title.font}{Title font of risk table (1 = normal, 2 = bold, 3 = italic, 4 = bold and italic).}

\item{risktable.title.col}{Colour for the risk table title. Can accept a single value for colour.}

\item{risktable.title.position}{Position of the title on the x-axis.}

\item{risktable.cex}{A numeric value specifying the size of the risk table size.}

\item{risktable.title.cex}{A numeric value specifying the size of the risk table title.}

\item{risktable.name.cex}{A numeric value specifying the size of the riksk legend name(s).}

\item{risktable.col}{Colour(s) for the risk table. Can accept a single value for colour, or a vector of colour values to set colour(s).}

\item{risktable.name.font}{legend name(s) font of risk table (1 = normal, 2 = bold, 3 = italic, 4 = bold and italic).}

\item{risktable.name.col}{Colour for the risk table name. Can accept a single value for colour.}

\item{risktable.name.position}{Position of the legend name(s) on the x-axis.}
}
\value{
Publication-Ready Kaplan-Meier Plot incorporating various statistics and layout customisation options to enhance the efficiency and adaptability of the Kaplan-Meier plot.
}
\description{
Plot publication ready Kaplan-Meier plot using the results from \code{survival::survfit()}.
}
\examples{
# Load Libraray
 require(survival)
 require(survSAKK)

# Load Data
 veteran$time_yr <- veteran$time/365.25
 veteran$time_mt <- veteran$time_yr*12

# Create survival object
 veteran_fit_yr <- survfit(Surv(time_yr, status) ~ 1, data = veteran)
 veteran_trt_fit_mt <- survfit(Surv(time_mt, status) ~ trt, data = veteran)

# Generate survival plots
 survSAKK::surv.plot(fit = veteran_fit_yr)

 survSAKK::surv.plot(
 fit = veteran_trt_fit_mt,
 col = c("#5ab4ac","#d8b365"),
 cex = 0.8,
 segment.quantile = 0.5,
 stat = "coxph",
 risktable.col = c("#5ab4ac","#d8b365"))

}
