% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survSAKK-package.R
\docType{package}
\name{survSAKK-package}
\alias{survSAKK}
\alias{survSAKK-package}
\title{survSAKK: Create publication ready Kaplan-Meier plot}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The survSAKK R package, incorporate various statistics and layout customization options to enhance the efficiency and adaptability of the Kaplan-Meier plot.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sakk-statistics.github.io/survSAKK/}
  \item Report bugs at \url{https://github.com/SAKK-Statistics/survSAKK/issues}
}

}
\author{
\strong{Maintainer}: Vithersan Somasundaram \email{Vithersan.Somasundaram@sakk.ch}

Authors:
\itemize{
  \item Katrin Gysel \email{Katrin.Gysel@sakk.ch} [reviewer]
}

Other contributors:
\itemize{
  \item Stefanie Hayoz \email{Stefanie.Hayoz@sakk.ch} [contributor]
  \item SAKK, CC [copyright holder, funder]
  \item SCTO, Statistics [funder]
}

}
\keyword{internal}
