% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.plot.R
\docType{package}
\name{surv.plot}
\alias{surv.plot}
\title{A package for publication ready Kaplan-Meier plots}
\usage{
surv.plot(
  fit,
  reference.arm,
  time.unit,
  y.unit = "probability",
  margin.bottom = 5,
  margin.left = NULL,
  margin.top = 3,
  margin.right = 2,
  censoring.mark = TRUE,
  censoring.cex = 1.3,
  conf.int = 0.95,
  conf.band = TRUE,
  conf.line = FALSE,
  conf.band.col = col,
  conf.band.transparent = 0.25,
  conf.type = "log-log",
  grid = FALSE,
  col = NULL,
  main = NULL,
  sub = NULL,
  xlab = NULL,
  ylab = NULL,
  xlab.pos = 1.5,
  ylab.pos = 3,
  xlab.cex = NULL,
  ylab.cex = NULL,
  cex = NULL,
  cex.axis,
  bty = "n",
  lty = c("solid", "dotted", "dotted"),
  lwd = 3,
  xticks,
  yticks = seq(from = 0, to = 1, by = 0.25),
  show.legend,
  legend.position = "topright",
  legend.name = NULL,
  legend.text.font = 1,
  legend.cex,
  legend.title = NULL,
  legend.title.cex,
  segment.type = 3,
  segment.timepoint = NULL,
  segment.quantile = NULL,
  segment.main = NULL,
  segment.annotation = "right",
  segment.col = "#666666",
  segment.annotation.col = col,
  segment.lty = "dotted",
  segment.lwd = 1.3,
  segment.cex,
  segment.confint = TRUE,
  segment.annotation.space = 0.06,
  segment.font = 1,
  segment.main.font = 1,
  stat = "none",
  stat.position = "bottomleft",
  stat.conf.int = 0.95,
  stat.fit,
  stat.col = "black",
  stat.cex,
  stat.font = 1,
  risktable = TRUE,
  risktable.pos = 2,
  risktable.title = "# at risk",
  risktable.title.font = 2,
  risktable.title.col = "black",
  risktable.title.position = par("usr")[1] - (par("usr")[2] - par("usr")[1]) * 0.15,
  risktable.name,
  risktable.cex,
  risktable.title.cex,
  risktable.name.cex,
  risktable.col = "black",
  risktable.name.font = 1,
  risktable.name.col = "black",
  risktable.name.position = par("usr")[1] - (par("usr")[2] - par("usr")[1]) * 0.15
)
}
\arguments{
\item{fit}{An object of class \link[survival:survfit]{survival::survfit} containing survival data.}

\item{reference.arm}{String defining the reference arm (optional)}

\item{time.unit}{The time unit of the survival curve.
Options:
\itemize{
\item \code{'day'}
\item \code{'week'}
\item \code{'month'}
\item \code{'year'}
}}

\item{y.unit}{Unit of the y-axis. Options: \code{'probability'}, \code{'percent'}}

\item{margin.bottom}{Specifies the bottom margin of the plotting area in line units. Default: 5}

\item{margin.left}{Specifies the left margin of the plotting area in line units. Default: 6 (with risktable) or 4 (without risktable)}

\item{margin.top}{Specifies the top margin of the plotting area in line units. Default: 3}

\item{margin.right}{Specifies the right margin of the plotting area in line units. Default: 2}

\item{censoring.mark}{A logical parameter indicating whether to mark censoring
events on the survival curves. Default: \code{TRUE}.}

\item{censoring.cex}{A numeric value specifying the size of the marks for
censored patients. Default: 1.3.}

\item{conf.int}{A numeric value between 0 and 1 controlling the confidence interval on the survival curves.
The default is 0.95 which corresponds to the 95\% confidence interval.
If set to 0, no confidence intervals are not displayed.
If a numeric value between 0 and 1 is provided, it represents the desired
coverage for the confidence interval (e.g. 0.9 for 90\%).}

\item{conf.band}{A logical parameter indicating whether to display the
confidence band on the survival curves. Default: \code{TRUE}.}

\item{conf.line}{A logical parameter indicating whether to draw the confidence
line on the survival curves. Default: \code{FALSE}.}

\item{conf.band.col}{Colour(s) for confidence band. Can accept a single value
for colour, or a vector of colour values.}

\item{conf.band.transparent}{A numeric values from 0 to 1. Controlling the
transparency of the confidence band. Default: 0.25.}

\item{conf.type}{Transformation type for the confidence interval.
\code{'log'}, \code{'log-log'} (default), \code{'plain'}, \code{'logit'}, \code{'arcsin'}.}

\item{grid}{A logical parameter specifying whether to draw a grid.
Default: \code{FALSE}.}

\item{col}{Colour(s) for the survival curves. Can accept a single value for
colour, or a vector of colour values to set colour(s).}

\item{main}{Title of the plot.}

\item{sub}{Subtitle of the plot. A subtitle only works if no risk table is displayed.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{xlab.pos}{Defines on which margin line the xlab is displayed. Starting at 0 counting outwards.}

\item{ylab.pos}{Defines on which margin line the ylab is displayed. Starting at 0 counting outwards.}

\item{xlab.cex}{A numeric value specifying the size of the X-axis label.}

\item{ylab.cex}{A numeric value specifying the size of the y-axis label.}

\item{cex}{A numeric value specifying all size of the text elements at once
(labels, annotations, ...).}

\item{cex.axis}{A numeric value specifying the size of the \code{axis} size.}

\item{bty}{Determines the style of the box drawn around the plot.
Options: \code{'n'} (default),\code{'o'},\code{'c'},\code{'u'}.}

\item{lty}{A vector with three arguments specifying line types for the curve
and the lower and upper confidence lines
Options for the three arguments: \code{'blank'}, \code{'solid'}, \code{'dashed'}, \code{'dotted'}, \code{'dotdash'},
\code{'longdash'}, \code{'twodash'}.
E.g. \code{c('solid', 'dashed', 'dashed')}.}

\item{lwd}{A numeric value specifying the width of the line.}

\item{xticks}{A numeric vector specifying the ticks of the x-axis. Can be specified as
\verb{seq(starting value, end value, number: increment of the sequence)}.}

\item{yticks}{A numeric vector specifying the ticks of the y-axis. Can be specified as
\verb{seq(starting value, end value, number: increment of the sequence)}.
It should always be specified as probability. For percent the parameter
\code{y.unit} can be used.}

\item{show.legend}{A logical parameter specifying whether to display legend.
By default the legend is displayed if there is more than one arm.}

\item{legend.position}{Position of the legend.
Options: \code{c(x,y)}, \code{'bottomright'}, \code{'bottom'}, \code{'bottomleft'}, \code{'left'},
'\verb{topleft'}, \code{'top'}, \code{'topright'}, \code{'right'}, \code{'center'}.}

\item{legend.name}{Renaming the name(s) of the arm.}

\item{legend.text.font}{Font style of the legend text.
\itemize{
\item \code{1} normal
\item \code{2} bold
\item \code{3} italic
\item \code{4} bold and italic
}}

\item{legend.cex}{A numeric value specifying the size of the legend text.}

\item{legend.title}{Title of the legend.}

\item{legend.title.cex}{A numeric value specifying the size of the legend title.}

\item{segment.type}{A numeric value specifying the layout of the segment.
\itemize{
\item \code{1} full width
\item \code{2} half width
\item \code{3} vertical and horizontal segment (default)
}}

\item{segment.timepoint}{A single value or a vector of fixed time points
to be drawn as segment(s).}

\item{segment.quantile}{A single value or a vector of fixed quantile to be
drawn as segment(s) e.g. 0.5 corresponds to median.}

\item{segment.main}{Title of segment text.}

\item{segment.annotation}{Position of the segment annotation.
Options: \code{c(x,y)},\code{'bottomleft'}, \code{'left'}, \code{'right'}, \code{'top'}, \code{'none'}.}

\item{segment.col}{Colour for the segment. Can accept a single value for colour.}

\item{segment.annotation.col}{Colour(s) for the segment annotation.
Can accept a single value for colour, or a vector of colour values to
set colour(s).}

\item{segment.lty}{A vector of string specifying line types for each curve.
Options: \code{'blank'}, \code{'solid'}, \code{'dashed'}, \code{'dotted'}, \code{'dotdash'},
\code{'longdash'}, \code{'twodash'}.}

\item{segment.lwd}{A numeric value specifying the width of the segment line.}

\item{segment.cex}{A numeric value specifying the size of the segment text size.}

\item{segment.confint}{Display the confidence interval.
\code{segment.confint} = \code{FALSE} is only possible if there are two arms.
Default: \code{TRUE}.}

\item{segment.annotation.space}{Spacing between the text in unit of x-coordinates.}

\item{segment.font}{A numeric value specifying the font face.
\itemize{
\item \code{1} plain
\item \code{2} bold
\item \code{3} italic
\item \code{4} bold-italic
}}

\item{segment.main.font}{A numeric value specifying the font face.
\itemize{
\item \code{1} plain
\item \code{2} bold
\item \code{3} italic
\item \code{4} bold-italic
}}

\item{stat}{Statistics which is displayed in the plot.
Options:
\itemize{
\item \code{'logrank'} gives the p value of the conducted logrank test using \verb{survdiff\{survival\}}.
To tests if there is a difference between two or more survival curves.
\item \code{'coxph'}  gives the hazard ratio (HR) and its CI (default: 95\% CI) of the conducted
Cox proportional hazards regression using \verb{coxph\{survival\}}. This option
only works if there are two arms.
\item \code{'coxph_logrank'}  combines the hazard ratio (HR), its CI (default: 95\% CI) and the
logrank test. This option only works if there are two arms.
\item \code{'coxmodel'} gives \code{N} (number of observations), \code{Events} (Number of events),
\code{HR}(hazard ratio), \code{lwrCI} (lower 95\% confidence interval),
\code{uprCI} (upper 95\% confidence interval) and \code{Logrank} (p-value corresponding to the Chisquare statistic)
of the conduct Cox proportional hazards regression using \verb{summary(coxph\{survival\})}.
\item \code{'none'} no statistic is displayed (default).
}}

\item{stat.position}{Position where the stat should be displayed.
Options: specify explicit by \code{c(x,y)},\code{'bottomleft'}, \code{'left'}, \code{'right'},
\code{'top'}, \code{'topright'},\code{'bottomright'}, \code{'none'}.}

\item{stat.conf.int}{Controlling the confidence interval on the hazard ratio.
If a numeric value between 0 and 1 is provided, it represents the desired
coverage for the confidence interval (e.g. 0.9 for 90\%).
Default: 0.95}

\item{stat.fit}{An second object of class \code{survfit} which is used for
calculation of statistics. This allows to add stratification factors.
Optional parameter.}

\item{stat.col}{Colour of the \code{stat} text. Can accept a single value for colour.}

\item{stat.cex}{A numeric value specifying the size of the \code{stat} text size.}

\item{stat.font}{The font face.
\itemize{
\item \code{1} plain
\item \code{2} bold
\item \code{3} italic
\item \code{4} bold-italic
}}

\item{risktable}{A logical parameter indicating whether to draw risk table. Default: \code{TRUE}.}

\item{risktable.pos}{Defines on which margin line the xlab is displayed. Starting at 0 counting outwards. Default at line 3.}

\item{risktable.title}{Title of risk table.}

\item{risktable.title.font}{Title font of risk table.
\itemize{
\item \code{1} normal
\item \code{2} bold
\item \code{3} italic
\item \code{4} bold and italic
}}

\item{risktable.title.col}{Colour for the risk table title. Can accept a single value for colour.}

\item{risktable.title.position}{A numeric value specifying the position of the title on the x-axis.}

\item{risktable.name}{Names of the arms for the risk table only.}

\item{risktable.cex}{A numeric value specifying the size of the risk table text size.}

\item{risktable.title.cex}{A numeric value specifying the size of the risk table title size.}

\item{risktable.name.cex}{A numeric value specifying the size of the risk table legend name size.}

\item{risktable.col}{Colour(s) for the risk table. Can accept a single value for colour, or a vector of colour values to set colour(s).
If it is set to \code{TRUE} then the colors of the curves are used.
Default: black}

\item{risktable.name.font}{legend name(s) font of risk table.
\itemize{
\item \code{1} normal
\item \code{2} bold
\item \code{3} italic
\item \code{4} bold and italic
}}

\item{risktable.name.col}{Colour for the risk table name. Can accept a single value for colour.}

\item{risktable.name.position}{A numeric value specifying the position of the legend name(s) on the x-axis.}
}
\value{
Publication-Ready Kaplan-Meier Plot incorporating various statistics and layout customisation options to enhance the efficiency and adaptability of the Kaplan-Meier plot.
}
\description{
The survSAKK R package provides the function \code{\link[=surv.plot]{surv.plot()}} to perform Kaplan-Meier survival analysis.
This package is designed to be user-friendly and efficient, offering robust tool for
analysing survival data and generating Kaplan-Meier plot using the results
from \code{\link[survival:survfit]{survival::survfit()}}.
}
\examples{
 require(survival)
 require(survSAKK)

# Create survival object
 veteran_trt_fit_mt <- survfit(Surv(time / 365.12 * 12, status) ~ trt, data = veteran)

# Generate survival plots
 surv.plot(fit = veteran_trt_fit_mt,
   time.unit = "month",
   reference.arm = 1,
   legend.name =  c("Standard", "Test"))

}
\references{
\code{vignette("surv.plot", package = "survSAKK")}
}
\seealso{
\itemize{
\item https://sakk-statistics.github.io/survSAKK/
\item \code{\link[survival:survfit]{survival::survfit()}} which this function wraps.
}
}
\author{
Vithersan Somasundaram and Katrin Gysel
}
