% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.plot.R
\name{surv.plot}
\alias{surv.plot}
\title{Publication Ready Kaplan-Meier Plot}
\usage{
surv.plot(
  fit,
  mark.censoring = TRUE,
  conf.int = fit$conf.int,
  conf.band = TRUE,
  conf.line = FALSE,
  conf.band.col = col,
  conf.band.transparent = 0.25,
  conf.type = "log-log",
  grid = FALSE,
  col = NULL,
  main = NULL,
  sub = NULL,
  xlab = "Time",
  ylab = "Estimated survival probability",
  cex = NULL,
  cex.lab = 1,
  cex.axis = 1,
  bty = "l",
  lty = c("solid", "dotted", "dotted"),
  lwd = 1,
  xlim = seq(from = 0, to = ceiling(max(fit$time)) + ceiling(min(fit$time))),
  ylim = seq(from = 0, to = 1, by = 0.25),
  show.legend = TRUE,
  legend.position = "topright",
  legend.name = NULL,
  legend.text.font = 1,
  legend.cex = 1,
  legend.title = NULL,
  legend.title.cex = 1,
  segment.type = 3,
  segment.timepoint = NULL,
  segment.quantile = NULL,
  segment.main = NULL,
  segment.annotation = "right",
  segment.col = "#666666",
  segment.annotation.col = col,
  segment.lty = "dashed",
  segment.lwd = 1,
  segment.cex = 1,
  segment.annotation.space = 0.06,
  segment.font = 1,
  segment.main.font = 1,
  stat = "none",
  stat.position = "bottomleft",
  stat.col = "black",
  stat.cex = 1,
  stat.font = 1,
  risktable = TRUE,
  risktable.axislab.pos = 2.5,
  risktable.margin.bottom = 5,
  risktable.margin.left = 6.5,
  risktable.title = "Number at risk",
  risktable.title.font = 2,
  risktable.title.col = "black",
  risktable.title.position = par("usr")[1] - (par("usr")[2] - par("usr")[1]) * 0.15,
  risktable.cex = 1,
  risktable.title.cex = 1,
  risktable.name.cex = 1,
  risktable.col = "black",
  risktable.name.font = 1,
  risktable.name.col = "#666666",
  risktable.name.position = par("usr")[1] - (par("usr")[2] - par("usr")[1]) * 0.15
)
}
\arguments{
\item{fit}{An object of class \code{survfit} containing survival data.}

\item{mark.censoring}{A logical parameter indicating whether to mark censoring
events on the survival curves. Default: \code{TRUE}.}

\item{conf.int}{Controlling the confidence interval on the survival curves.
When set to \code{TRUE}, the function displays default 95\% confidence interval.
If set to \code{FALSE}, confidence intervals are not displayed.
If a numeric value between 0 and 1 is provided, it represents the desired
coverage for the confidence interval (e.g. 0.9 for 90\%).}

\item{conf.band}{A logical parameter indicating whether to display the
confidence band on the survival curves. Default: \code{TRUE}.}

\item{conf.line}{A logical parameter indicating whether to draw the confidence
line on the survival curves. Default: \code{FALSE}.}

\item{conf.band.col}{Colour(s) for confidence band. Can accept a single value
for colour, or a vector of colour values.}

\item{conf.band.transparent}{A numeric values from 0 to 1. Controlling the
transparency of the confidence band. Default: 0.25.}

\item{conf.type}{Transformation type for the confidence interval.
\code{'log'}, \code{'log-log'} (default), \code{'plain'}, \code{'logit'}, \code{'arcsin'}.}

\item{grid}{A logical parameter specifying whether to draw a grid.
Default: \code{FALSE}.}

\item{col}{Colour(s) for the survival curves. Can accept a single value for
colour, or a vector of colour values to set colour(s).}

\item{main}{Title of the plot.}

\item{sub}{Subtitle of the plot.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{cex}{A numeric value specifying all size of the text elements at once
(labels, annotations, ...).}

\item{cex.lab}{A numeric value specifying the size of the \code{xlab} and \code{ylab} text.}

\item{cex.axis}{A numeric value specifying the size of the \code{axis} size.}

\item{bty}{Determines the style of the box drawn around the plot.
Options: \code{'n'},\code{'o'},\code{'7'},\code{'L'},\code{'C'},\code{'U'}.}

\item{lty}{A vector of string specifying line types for each curve.
The length of the vector should match the number of survival curves,
assigning a specific line type to each curve. in the plot.
Options: \code{'blank'}, \code{'solid'}, \code{'dashed'}, \code{'dotted'}, \code{'dotdash'},
\code{'longdash'}, \code{'twodash'}.}

\item{lwd}{A numeric value specifying the width of the line.}

\item{xlim}{Limits for the x-axis. Specified as
\verb{seq(starting value, end value, number of increment of the sequence)}.}

\item{ylim}{Limits for the y-axis. Specified as
\verb{seq(starting value, end value, number of increment of the sequence)}.}

\item{show.legend}{A logical parameter specifying whether to display legend.
Default: \code{TRUE}.}

\item{legend.position}{Position of the legend.
Options: \code{c(x,y)}, \code{'bottomright'}, \code{'bottom'}, \code{'bottomleft'}, '\code{left}',
'\verb{topleft'}, \code{'top'}, \code{'topright'}, \code{'right'}, \code{'center'}.}

\item{legend.name}{Renaming the name(s) of the stratum.}

\item{legend.text.font}{Font style of the legend text.
\itemize{
\item \code{1} normal
\item \code{2} bold
\item \code{3} italic
\item \code{4} bold and italic
}}

\item{legend.cex}{A numeric value specifying the size of the legend text.}

\item{legend.title}{Title of the legend.}

\item{legend.title.cex}{A numeric value specifying the size of the legend title.}

\item{segment.type}{A numeric value specifying the layout of the segment.
\itemize{
\item \code{1} (full width)
\item \code{2} (half width)
\item \code{3} (vertical and horizontal segment)
}}

\item{segment.timepoint}{A single value or a vector of fixed time points
to be drawn as segment(s).}

\item{segment.quantile}{A single value or a vector of fixed quantile to be
drawn as segment(s) e.g. 0.5 corresponds to median.}

\item{segment.main}{Title of segment text.}

\item{segment.annotation}{Position of the segment annotation.
Options: \code{c(x,y)},\code{'bottomleft'}, \code{'left'}, \code{'right'}, \code{'none'}.}

\item{segment.col}{Colour for the segment.  Can accept a single value for colour.}

\item{segment.annotation.col}{Colour(s) for the segment annotation.
Can accept a single value for colour, or a vector of colour values to
set colour(s).}

\item{segment.lty}{A vector of string specifying line types for each curve.
Options: \code{'blank'}, \code{'solid'}, \code{'dashed'}, \code{'dotted'}, \code{'dotdash'},
\code{'longdash'}, \code{'twodash'}.}

\item{segment.lwd}{A numeric value specifying the width of the segment line.}

\item{segment.cex}{A numeric value specifying the size of the segment text size.}

\item{segment.annotation.space}{Spacing between the text in unit of x-coordinates.}

\item{segment.font}{A numeric value specifying the font face.
\itemize{
\item \code{1} plain
\item \code{2} bold
\item \code{3} italic
\item \code{4} bold-italic
}}

\item{segment.main.font}{A numeric value specifying the font face.
\itemize{
\item \code{1} plain
\item \code{2} bold
\item \code{3} italic
\item \code{4} bold-italic
}}

\item{stat}{Statistics which is displayed in the plot.
Options:
\itemize{
\item \code{'logrank'} gives the p value of the conducted logrank test using \verb{survdiff\{survival\}}.
To tests if there is a difference between two or more survival curves.
\item \code{'coxph'}  gives the hazard ratio (HR) and its 95\% CI of the conducted
Cox proportional hazards regression using \verb{coxph\{survival\}}.
\item \code{'coxmodel'} gives \code{N} (number of observations), \code{Events} (Number of events),
\code{HR}(hazard ratio), \code{lwrCI} (lower 95\% confidence interval),
\code{uprCI} (upper 95\% confidence interval) and \code{Logrank} (p-value corresponding to the Chisquare statistic)
of the conduct Cox proportional hazards regression using \verb{summary(coxph\{survival\})}.
\item \code{'none'} no statistic is displayed (default).
}}

\item{stat.position}{Position where the stat should be displayed.
Options: specify explicit by \code{c(x,y)},\code{'bottomleft'}, \code{'left'}, \code{'right'}, \code{'topright'},\code{'bottomright'}, \code{'none'}.}

\item{stat.col}{Colour of the \code{stat} text. Can accept a single value for colour.}

\item{stat.cex}{A numeric value specifying the size of the \code{stat} text size.}

\item{stat.font}{The font face.
\itemize{
\item \code{1} plain
\item \code{2} bold
\item \code{3} italic
\item \code{4} bold-italic
}}

\item{risktable}{A logical parameter indicating whether to draw risk table. Default: \code{TRUE}.}

\item{risktable.axislab.pos}{Specifies on which line on the plotting area \code{X} and \code{Y} label
should be drawn if \code{risktable} is drawn. Default: 2.5 line distances form the axis elements.}

\item{risktable.margin.bottom}{Specifies the bottom margin of the plotting area for the \code{risktable}in line units.
Bottom margin line is calculated by \code{risktable.margin.bottom} (default: 5 line units + Number of stratum).}

\item{risktable.margin.left}{Specifies the left margin of the plotting area for the \code{risktable} in line units.
Left margin line default is set by 6.5 line units.}

\item{risktable.title}{Title of risk table.}

\item{risktable.title.font}{Title font of risk table.
\itemize{
\item \code{1} normal
\item \code{2} bold
\item \code{3} italic
\item \code{4} bold and italic
}}

\item{risktable.title.col}{Colour for the risk table title. Can accept a single value for colour.}

\item{risktable.title.position}{A numeric value specifying the position of the title on the x-axis.}

\item{risktable.cex}{A numeric value specifying the size of the risk table text size.}

\item{risktable.title.cex}{A numeric value specifying the size of the risk table title size.}

\item{risktable.name.cex}{A numeric value specifying the size of the rsik table legend name size.}

\item{risktable.col}{Colour(s) for the risk table. Can accept a single value for colour, or a vector of colour values to set colour(s).}

\item{risktable.name.font}{legend name(s) font of risk table.
\itemize{
\item \code{1} normal
\item \code{2} bold
\item \code{3} italic
\item \code{4} bold and italic
}}

\item{risktable.name.col}{Colour for the risk table name. Can accept a single value for colour.}

\item{risktable.name.position}{A numeric value specifying the position of the legend name(s) on the x-axis.}
}
\value{
Publication-Ready Kaplan-Meier Plot incorporating various statistics and layout customisation options to enhance the efficiency and adaptability of the Kaplan-Meier plot.
}
\description{
Plot publication ready Kaplan-Meier plot using the results from \code{survival::survfit()}.
}
\examples{
# Load Libraray
 require(survival)
 require(survSAKK)

# Load Data
 veteran$time_yr <- veteran$time/365.25
 veteran$time_mt <- veteran$time_yr*12

# Create survival object
 veteran_fit_yr <- survfit(Surv(time / 365.25, status) ~ 1, data = veteran)
 veteran_trt_fit_mt <- survfit(Surv(time / 365.12 * 12, status) ~ trt, data = veteran)

# Generate survival plots
 survSAKK::surv.plot(fit = veteran_fit_yr)

 survSAKK::surv.plot(
 fit = veteran_trt_fit_mt,
 col = c("#5ab4ac","#d8b365"),
 cex = 0.8,
 segment.quantile = 0.5,
 stat = "coxph",
 risktable.col = c("#5ab4ac","#d8b365"))

}
\references{
\code{vignette("surv.plot", package = "survSAKK")}
}
